+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/SchoolRestController.java
@@ -39,4 +39,9 @@ public class SchoolRestController {
     public void deleteSchool(@PathVariable Long schoolId){
         schoolController.deleteSchool(schoolId);
     }
+
+    @GetMapping("/centerName/{schoolId}")
+    public String getCenterNameBySchoolId(@PathVariable Long schoolId){
+        return schoolController.getCenterNameBySchoolId(schoolId);
+    }
 }