+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/SchoolController.java
@@ -60,4 +60,8 @@ public class SchoolController {
         schoolService.deleteSchool(schoolId);
     }
 
+    public String getCenterNameBySchoolId(Long schoolId) {
+        School school = schoolService.getSchoolById(schoolId);
+        return school.getCenter().getCenterName();
+    }
 }