+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -45,3 +45,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [NPP-44] 》ST38F - Create a student associated with a course
 - [NPP-55] 》ST46F - Create a student without associated course
+
+- [NPP-56] 》EX47 - Get center name from school by schoolID
+